UI Plugin: Move and rename some the MC create workflow components
Moves the AWS and Docker MC create workflows into the views folder

Renamed some of these files to be more representative of their purpose, but open to alternative suggestions
