+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/management-cluster/aws/AwsManagementCluster.scss
@@ -1,7 +1,7 @@
 .aws-management-container {
     .aws-sm-logo {
         display: inline-block;
-        background-image: url('../../assets/aws.svg');
+        background-image: url('../../../assets/aws.svg');
         background-repeat: no-repeat;
         background-position: top right;
         width: 48px;
