+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/management-cluster/aws/wizard-simple/management-credential-step/ManagementCredentialOneTime.scss
