+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/management-cluster/docker/DockerManagementCluster.scss
@@ -2,7 +2,7 @@
     h2 {
         .docker-sm-logo {
             display: inline-block;
-            background-image: url('../../assets/docker.svg');
+            background-image: url('../../../assets/docker.svg');
             background-repeat: no-repeat;
             background-position: top right;
             width: 48px;
