+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/management-cluster/docker/DockerManagementCluster.tsx
@@ -2,33 +2,33 @@
 import React from 'react';
 
 // App imports
-import { AwsProvider } from '../../state-management/stores/Store.aws';
-import AwsMCCreateSimple from '../../components/aws/AwsMCCreateSimple';
-import './AwsManagementClusterSimple.scss';
-import AwsLogo from '../../assets/aws.svg';
+import { DockerProvider } from '../../../state-management/stores/Docker.store';
+import DockerManagementClusterSimple from './wizard-simple/DockerManagementClusterSimple';
+import DockerLogo from '../../../assets/docker.svg';
+import './DockerManagementCluster.scss';
 
-function AwsManagementClusterSimple() {
+function DockerManagementCluster() {
     return (
-        <AwsProvider>
+        <DockerProvider>
             <div className="aws-management-container" cds-layout="grid col:12">
                 <div cds-layout="col:8">
                     <div cds-layout="col:12 p-b:lg">
                         <span cds-text="title">
-                            <img src={AwsLogo} className="logo logo-42" cds-layout="m-r:md" alt="aws logo" />
-                            Create Management Cluster on Amazon Web Services
+                            <img src={DockerLogo} className="logo logo-42" cds-layout="m-r:md" alt="aws logo" />
+                            Create Management Cluster on Docker
                         </span>
                     </div>
 
                     {/* Disable Simple/Advanced banner until advanced settings available */}
                     {/* <ConfigBanner /> */}
-                    <AwsMCCreateSimple />
+                    <DockerManagementClusterSimple />
                 </div>
                 <div cds-layout="col:4" className="image-container">
                     <div className="mgmt-cluster-admins-img"></div>
                 </div>
             </div>
-        </AwsProvider>
+        </DockerProvider>
     );
 }
 
-export default AwsManagementClusterSimple;
+export default DockerManagementCluster;
