+++ cli/cmd/plugin/ui/web/tanzu-ui/src/views/management-cluster/docker/wizard-simple/DockerManagementClusterSimple.tsx
@@ -2,14 +2,14 @@
 import React, { useContext } from 'react';
 
 // App imports
-import { DockerStore } from '../../state-management/stores/Docker.store';
-import { TAB_NAMES } from '../../shared/constants/NavRoutes.constants';
-import Wizard from '../../shared/components/wizard/Wizard';
+import { DockerStore } from '../../../../state-management/stores/Docker.store';
+import { TAB_NAMES } from '../../../../shared/constants/NavRoutes.constants';
+import Wizard from '../../../../shared/components/wizard/Wizard';
 import McPrerequisiteStep from './mc-prerequisite-step/McPrerequisiteStep';
-import ManagementClusterSettings from '../../shared/components/management-cluster-settings-step/ManagementClusterSettings';
-import useDockerDeployment from '../../shared/services/dockerDeployment';
+import ManagementClusterSettings from '../../../../shared/components/management-cluster-settings-step/ManagementClusterSettings';
+import useDockerDeployment from '../../../../shared/services/dockerDeployment';
 
-function DockerMCCreateSimple() {
+function DockerManagementClusterSimple() {
     const { dockerState, dockerDispatch } = useContext(DockerStore);
     const { deployOnDocker } = useDockerDeployment();
     return (
@@ -25,4 +25,4 @@ function DockerMCCreateSimple() {
     );
 }
 
-export default DockerMCCreateSimple;
+export default DockerManagementClusterSimple;
